+++ src/deployment-model/contracts/environment-validation-result.js
@@ -1,0 +1,8 @@
+export class EnvironmentValidationResult {
+  constructor({ isValid, errors = [] }) {
+    this.isValid = Boolean(isValid);
+    this.errors = Object.freeze([...errors]);
+
+    Object.freeze(this);
+  }
+}