+++ src/deployment-model/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { EnvironmentStrategyDescriptor } from './environments/environment-strategy-descriptor.js';
 import { DeploymentOverviewDescriptor } from './overview/deployment-overview-descriptor.js';
 
 export function addDeploymentModel(services) {
@@ -7,6 +8,7 @@ export function addDeploymentModel(services) {
   }
 
   services.registerSingleton('DeploymentOverviewDescriptor', () => new DeploymentOverviewDescriptor());
+  services.registerSingleton('EnvironmentStrategyDescriptor', () => new EnvironmentStrategyDescriptor());
 
   return services;
 }